memoire editAction

+++ src/Mondofute/Bundle/HebergementBundle/Controller/HebergementUnifieController.php
@@ -1580,7 +1580,10 @@ class HebergementUnifieController extends Controller
                         }
 
                         echo memory_get_usage() . PHP_EOL;
-                        die;
+                        if((memory_get_usage()/1024/1024)>= (intval(ini_get('memory_limit'),10)*60/100)){
+                            die;
+                        }
+//                        die;
                         // *** fin suppression des code promo logement ***
                         $em->remove($originalFournisseurHebergement);
                     }
